requirements.txt
aiogram>=3.0.0
python-dotenv
sqlalchemy>=2.0.0
aiosqlite 